Refactor: make FunctionalTransitionSystem inherit from RelationalTransitionSystem (compiles but untested)
